[Refactor] Move notification pages to own NotificationController (#3820)
* NotificationController refactor

* renderJsonSuccess & renderJsonFail

* renderJsonSuccess & renderJsonFail

* typo: notifcation -> notification

* add an else clause to handle exception & format changes
