+++ site/app/templates/NotificationSettings.twig
@@ -4,7 +4,7 @@
     }
 </style>
 <div class="content">
-    <form method="post" action="{{ core.buildUrl({'component': 'notification_settings', 'page': 'alter_notifcation_settings'}) }}" id="form_notification_settings">
+    <form method="post" action="{{ core.buildUrl({'component': 'notification', 'page': 'alter_notification_settings'}) }}" id="form_notification_settings">
         <div id="config">
             <div><h1>Notification/Email Settings</h1></div>
             <div class="row">
@@ -136,10 +136,10 @@
                 success: function(data) {
                     try {
                         var json = JSON.parse(data);
-                        if(json['error']) {
-                            var message ='<div class="inner-message alert alert-error" style="position: fixed;top: 40px;left: 50%;width: 40%;margin-left: -20%;" id="theid"><a class="fas fa-times message-close" onclick="removeMessagePopup(\'theid\');"></a><i class="fas fa-times-circle"></i>' + json['error'] + '</div>';
+                        if(json['status'] == 'fail') {
+                            var message ='<div class="inner-message alert alert-error" style="position: fixed;top: 40px;left: 50%;width: 40%;margin-left: -20%;" id="theid"><a class="fas fa-times message-close" onclick="removeMessagePopup(\'theid\');"></a><i class="fas fa-times-circle"></i>' + json['message'] + '</div>';
                         }else {
-                            var message ='<div class="inner-message alert alert-success" style="position: fixed;top: 40px;left: 50%;width: 40%;margin-left: -20%;" id="theid"><a class="fas fa-times message-close" onclick="removeMessagePopup(\'theid\');"></a><i class="fas fa-times-circle"></i>' + json['success'] + '</div>';
+                            var message ='<div class="inner-message alert alert-success" style="position: fixed;top: 40px;left: 50%;width: 40%;margin-left: -20%;" id="theid"><a class="fas fa-times message-close" onclick="removeMessagePopup(\'theid\');"></a><i class="fas fa-times-circle"></i>' + json['data'] + '</div>';
                         }
                     } catch(err) {
                         var message ='<div class="inner-message alert alert-error" style="position: fixed;top: 40px;left: 50%;width: 40%;margin-left: -20%;" id="theid"><a class="fas fa-times message-close" onclick="removeMessagePopup(\'theid\');"></a><i class="fas fa-times-circle"></i>Error parsing data. Please try again.</div>';
