+++ site/app/templates/Notifications.twig
@@ -6,7 +6,7 @@
         {% else %}
             <a href="{{ core.buildUrl({"component": "navigation", "page": "notifications", "show_all": "1"}) }}" class="btn btn-primary">Show All</a>
         {% endif %}
-        <a href="{{ core.buildUrl({"component": "navigation", "page": "notification_settings" }) }}" class="btn btn-primary">Settings</a>
+        <a href="{{ core.buildUrl({"component": "notification", "page": "notification_settings" }) }}" class="btn btn-primary">Settings</a>
     </div>
     <h1>Notifications</h1>
     {# This is a layout table #}
